Let RSpec manage load paths for us
RSpec will automatically add the lib/ to the load path, so in our spec_helper, we don't need to explicitly add the lib/.

RSpec will also add the directory we've told it to look for specifications in to the load path, so we can also omit the line that adds the spec/ to the load path.

And then in our spec/*_spec.rb files, we can just say `require 'spec_helper'` since spec/ is in the load path.