Add lib to the load path.
To make paths relative in lib, we need to have lib added to the load path
for any way we could potentially load it. Since one of these ways is through
the binary, we need to do it there.

Then, we can just require 'blirb' rather than requiring the absolute path to it.