Switch require paths to be relative.
Rubygems will automatically add the lib to the load path, because
we told it to do this in the gemspec (s.require_paths = ["lib"])

So, every way of invoking this lib (from the binary, from the
rspec suite, as a rubygem) already has the lib/ in the load path.
This means that we can use all relative paths within the lib.